Teknofestyazilim/requirements.txt
keras==2.6.0
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
scikit-image==0.17.2
scipy==1.5.4
sniffio==1.2.0
tensorflow==2.6.2
